migrate tokenomics to 18 decimals and BigInt arithmetic
Major economic refactor to ensure full EVM and MetaMask compatibility.

1. Updated BaseUnit from 1e9 to 1e18 (Wei standard).
   - This aligns Thrylos with standard Ethereum tooling.

2. Migrated Account and Transaction models from int64 to string.
   - int64 overflows at ~9 tokens when using 18 decimals.
   - Switched to string-based storage in Protobuf/JSON to support
     arbitrary precision without overflow or floating-point errors.

3. Implemented core/math/big_utils.go.
   - Added helper functions for BigInt arithmetic (Add, Sub, Mul).
   - Added safe String-to-BigInt parsing logic.

4. Updated Configuration and Genesis.
   - Config structs now use strings for all monetary fields.
   - Updated genesis loading logic to parse high-precision balances.

Further work needed to update the cofig

+++ core/block/creator.go
@@ -124,7 +124,8 @@ func (bc *Creator) ValidateBlockSize(block *core.Block) error {
 		return fmt.Errorf("failed to calculate block size: %v", err)
 	}
 
-	if blockSize > config.MaxBlockSize {
+	// FIX: Explicitly cast blockSize to int64
+	if int64(blockSize) > config.MaxBlockSize {
 		return fmt.Errorf("block too large: %d bytes > %d bytes limit", blockSize, config.MaxBlockSize)
 	}
